AiDetector/inference-server/pyproject.toml
[project]
name = "inference-server"
version = "0.1.0"
description = ""
authors = [
    {name = "Martin Prochazka",email = "pro.nitram@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "grpcio (>=1.75.1,<2.0.0)",
    "grpcio-tools (>=1.75.1,<2.0.0)",
    "grpcio-reflection (>=1.75.1,<2.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
]

[tool.poetry]
packages = [{include = "inference_server", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (==0.13.2)"
]

[tool.ruff]
extend-exclude = ["src/inference_server/proto"]